+++ CavrnusSdk/API/CavrnusLocalUser.cs
@@ -1,35 +1,35 @@
-using CavrnusSdk;
 using Collab.Base.Collections;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-public class CavrnusLocalUser : MonoBehaviour
+namespace CavrnusSdk
 {
-    void Start()
-    {
-		CavrnusSpaceJoinEvent.OnAnySpaceConnection(spaceConn =>
-		{
-			var user = spaceConn.UsersList.LocalUser.Bind(lu =>
+	public class CavrnusLocalUser : MonoBehaviour
+	{
+	    private void Start()
+	    {
+			CavrnusSpaceJoinEvent.OnAnySpaceConnection(spaceConn =>
 			{
-				gameObject.AddComponent<CavrnusUserFlag>().User = lu;
+				var user = spaceConn.UsersList.LocalUser.Bind(lu =>
+				{
+					gameObject.AddComponent<CavrnusUserFlag>().User = lu;
 
-				foreach (var propContainer in gameObject.GetComponentsInChildren<CavrnusPropertiesContainer>())
-					propContainer.PrefixContainerName($"{lu.ContainerId}");
+					foreach (var propContainer in gameObject.GetComponentsInChildren<CavrnusPropertiesContainer>())
+						propContainer.PrefixContainerName($"{lu.ContainerId}");
 
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<bool>>())
-					sync.RecieveChanges = false;
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<float>>())
-					sync.RecieveChanges = false;
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Color>>())
-					sync.RecieveChanges = false;
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Vector4>>())
-					sync.RecieveChanges = false;
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<CavrnusPropertyHelpers.TransformData>>())
-					sync.RecieveChanges = false;
-				foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<string>>())
-					sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<bool>>())
+						sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<float>>())
+						sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Color>>())
+						sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Vector4>>())
+						sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<CavrnusPropertyHelpers.TransformData>>())
+						sync.RecieveChanges = false;
+					foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<string>>())
+						sync.RecieveChanges = false;
+				});
 			});
-		});
+		}
 	}
 }